requirements.txt
dj-rest-auth==5.0.2
Django==4.2.7
djangorestframework==3.14.0
psycopg2==2.9.9

django-allauth==0.58.2
djangorestframework-simplejwt==5.3.0
rest-framework-simplejwt==0.0.2

